qt-cpp: Replace natvis preparation script with simple copy
The debugging_helpers submodule no longer contains ##NAMESPACE##
markers in natvis files, so the marker-stripping logic and timestamp
caching in prepare_natvis_files.ts are obsolete. Replace it with a
minimal copy_natvis_files.ts that just copies the files as-is.
